+++ src/quantum/azext_quantum/_help.py
@@ -381,27 +381,35 @@ helps['quantum workspace user list'] = """
             az quantum workspace user list -g MyResourceGroup -w MyWorkspace --include-inherited false
 """
 
-helps['quantum workspace user create'] = """
+helps['quantum workspace user add'] = """
     type: command
     short-summary: Grant a user access to an Azure Quantum workspace.
     long-summary: >-
         Assigns the 'Quantum Workspace Data Contributor' role at the scope of the given
         (or current) Azure Quantum workspace.
     examples:
-      - name: Grant a user access to a workspace using their sign-in name (email).
+      - name: Grant a user access to a workspace using their principal name.
         text: |-
-            az quantum workspace user create -g MyResourceGroup -w MyWorkspace \\
-                --email user@contoso.com
+            az quantum workspace user add -g MyResourceGroup -w MyWorkspace \\
+                --user user@contoso.com
+      - name: Grant a user access to a workspace using their object ID.
+        text: |-
+            az quantum workspace user add -g MyResourceGroup -w MyWorkspace \\
+                --user 00000000-0000-0000-0000-000000000000
 """
 
-helps['quantum workspace user delete'] = """
+helps['quantum workspace user remove'] = """
     type: command
     short-summary: Remove a user's access to an Azure Quantum workspace.
     examples:
-      - name: Remove a user's access to a workspace using their sign-in name (email).
+      - name: Remove a user's access to a workspace using their principal name.
         text: |-
-            az quantum workspace user delete -g MyResourceGroup -w MyWorkspace \\
-                --email user@contoso.com
+            az quantum workspace user remove -g MyResourceGroup -w MyWorkspace \\
+                --user user@contoso.com
+      - name: Remove a user's access to a workspace using their object ID.
+        text: |-
+            az quantum workspace user remove -g MyResourceGroup -w MyWorkspace \\
+                --user 00000000-0000-0000-0000-000000000000
 """
 
 helps['quantum workspace keys list'] = """